moved login to component
